requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.6.0
multidict==6.4.3
yarl==1.20.0
python-dotenv==1.1.0
librosa==0.11.0
numpy==1.26.4
soundfile==0.13.1
soxr==0.5.0.post1
torch==2.2.0
